+++ src/event-sources/index.js
@@ -2,6 +2,8 @@ const awsApiGatewayV1EventSource = require('./aws/api-gateway-v1')
 const awsApiGatewayV2EventSource = require('./aws/api-gateway-v2')
 const awsAlbEventSource = require('./aws/alb')
 const awsLambdaEdgeEventSource = require('./aws/lambda-edge')
+const awsSnsEventSource = require('./aws/sns')
+const awsDynamoDbEventSource = require('./aws/dynamodb')
 
 function getEventSource ({ eventSourceName }) {
   switch (eventSourceName) {
@@ -13,6 +15,10 @@ function getEventSource ({ eventSourceName }) {
       return awsAlbEventSource
     case 'AWS_LAMBDA_EDGE':
       return awsLambdaEdgeEventSource
+    case 'AWS_DYNAMODB':
+      return awsDynamoDbEventSource
+    case 'AWS_SNS':
+      return awsSnsEventSource
     default:
       throw new Error('Couldn\'t detect valid event source.')
   }